requirements.txt
Flask==3.0.2
Pillow==10.2.0
pyautogui==0.9.54
python-dotenv==1.0.1
easyocr==1.7.1
google-generativeai==0.3.2
requests==2.31.0
pygame==2.5.2 